Update WPF viewer
