+++ src/WPF/WPF.Viewer/Converters/DirectionPointTypeToIconConverter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Windows.Data;
 
-namespace ArcGIS.Converters
+namespace ArcGIS.WPF.Viewer.Converters
 {
     internal class DirectionPointTypeToIconConverter : IValueConverter
     {
