+++ src/WPF/WPF.Viewer/Converters/TreeViewLeftMarginMultiplierConverter.cs
@@ -5,7 +5,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Data;
 
-namespace ArcGIS.Converters
+namespace ArcGIS.WPF.Viewer.Converters
 {
     // Code taken from Stack Overflow answer: https://stackoverflow.com/a/672123/4630559
     public class LeftMarginMultiplierConverter : IValueConverter
