+++ src/WPF/WPF.Viewer/Samples/Layers/DisplayRouteLayer/DisplayRouteLayer.xaml
@@ -1,43 +1,43 @@
 ﻿<UserControl x:Class="ArcGIS.WPF.Samples.DisplayRouteLayer.DisplayRouteLayer"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:converters="clr-namespace:ArcGIS.Converters"
+             xmlns:converters="clr-namespace:ArcGIS.WPF.Viewer.Converters"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
-    <UserControl.Resources>
-        <converters:DirectionPointTypeToIconConverter x:Key="Converter" />
-    </UserControl.Resources>
-    <Grid>
-        <esri:MapView x:Name="MyMapView" />
-        <Border Height="300" Style="{StaticResource BorderStyle}">
-            <ScrollViewer Height="250" VerticalScrollBarVisibility="Auto">
-                <StackPanel>
-                    <TextBlock FontWeight="Bold" Text="Directions:" />
-                    <ItemsControl x:Name="DirectionsList">
-                        <ItemsControl.ItemTemplate>
-                            <DataTemplate>
-                                <Grid>
-                                    <Grid.ColumnDefinitions>
-                                        <ColumnDefinition Width="20" />
-                                        <ColumnDefinition Width="25" />
-                                        <ColumnDefinition Width="*" />
-                                    </Grid.ColumnDefinitions>
-                                    <TextBlock Margin="0,0,5,5"
-                                               HorizontalAlignment="Right"
-                                               Text="{Binding Attributes[Sequence], StringFormat={}{0}.}" />
-                                    <TextBlock Grid.Column="1"
-                                               Margin="0,0,5,5"
-                                               FontFamily="pack://application:,,,/Resources/Fonts/#calcite-ui-icons-24"
-                                               FontSize="20"
-                                               Text="{Binding Attributes[DirectionPointType], Converter={StaticResource Converter}}" />
-                                    <TextBlock Grid.Column="2"
-                                               Text="{Binding Attributes[DisplayText]}"
-                                               TextWrapping="Wrap" />
-                                </Grid>
-                            </DataTemplate>
-                        </ItemsControl.ItemTemplate>
-                    </ItemsControl>
-                </StackPanel>
-            </ScrollViewer>
-        </Border>
-    </Grid>
+	<UserControl.Resources>
+		<converters:DirectionPointTypeToIconConverter x:Key="Converter" />
+	</UserControl.Resources>
+	<Grid>
+		<esri:MapView x:Name="MyMapView" />
+		<Border Height="300" Style="{StaticResource BorderStyle}">
+			<ScrollViewer Height="250" VerticalScrollBarVisibility="Auto">
+				<StackPanel>
+					<TextBlock FontWeight="Bold" Text="Directions:" />
+					<ItemsControl x:Name="DirectionsList">
+						<ItemsControl.ItemTemplate>
+							<DataTemplate>
+								<Grid>
+									<Grid.ColumnDefinitions>
+										<ColumnDefinition Width="20" />
+										<ColumnDefinition Width="25" />
+										<ColumnDefinition Width="*" />
+									</Grid.ColumnDefinitions>
+									<TextBlock Margin="0,0,5,5"
+									           HorizontalAlignment="Right"
+									           Text="{Binding Attributes[Sequence], StringFormat={}{0}.}" />
+									<TextBlock Grid.Column="1"
+									           Margin="0,0,5,5"
+									           FontFamily="pack://application:,,,/Resources/Fonts/#calcite-ui-icons-24"
+									           FontSize="20"
+									           Text="{Binding Attributes[DirectionPointType], Converter={StaticResource Converter}}" />
+									<TextBlock Grid.Column="2"
+									           Text="{Binding Attributes[DisplayText]}"
+									           TextWrapping="Wrap" />
+								</Grid>
+							</DataTemplate>
+						</ItemsControl.ItemTemplate>
+					</ItemsControl>
+				</StackPanel>
+			</ScrollViewer>
+		</Border>
+	</Grid>
 </UserControl>
